use lodash debounce and mock it
