--- packages/react/src/utils/debounce.ts
@@ -1,23 +1,0 @@
-const debounce = (fn: Function, time: number): Function => {
-  let timeoutId
-
-  function cancel() {
-    if (timeoutId) {
-      clearTimeout(timeoutId)
-    }
-  }
-
-  function wrapper(...args) {
-    cancel()
-    timeoutId = setTimeout(() => {
-      timeoutId = null
-      fn(...args)
-    }, time)
-  }
-
-  wrapper.cancel = cancel
-
-  return wrapper
-}
-
-export default debounce
