+++ packages/react/src/utils/index.ts
@@ -49,4 +49,3 @@ export module commonPropTypes {
 
 export { default as withDebugId } from './withDebugId'
 export { default as Telemetry } from './Telemetry'
-export { default as debounce } from './debounce'
